requirements.txt
python-dotenv
google-generativeai
fastapi
uvicorn
pandas
